Add "port connected" component class method
What used to be "accept port connection" is now split in "accept port
connection" and "port connected". The latter confirms that the two ports
are connected after they both accepted the connection. Thus in the
"accept port connection" method, the ports are not connected yet: this
is not where the component can create a notification iterator because
the connection object does not exist.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>