Event notification: validate that CC in the CC prio map have a value
Instead of making sure that all the trace's clock classes have a value
and a priority in an event, instead make sure that all the clock classes
found in the event notification's clock class priority map have a value
in the event and that this clock class is part of the event's trace.

This makes it possible for an event to have only clock values for
relevant clock classes, not all the trace's, or even to have no value
at all.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>