Fix: ir: bt_ctf_field_type_variant_validate(): do not compare to tag length
Because the same label can map to multiple values in an enumeration FT,
the number of fields in a variant FT does not need to be the same as the
number of mappings in its tag FT for the variant FT to be considered
valid.

Relax this by only making sure that each tag FT mapping has an entry in
the variant FT which shares the same name. Unused variant FT
fields/choices are not a problem per se, just a waste of space.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>