Add graph cancellation API
With this API you can cancel a graph with bt_graph_cancel() and
check if a graph is canceled with bt_graph_is_canceled().

A canceled graph is used to indicate to a contained component or
notification iterator if it should retry or not when a system call is
interrupted. This exists mostly to support a clean termination when we
get SIGINT, but with debugger support (a system call is interrupted when
the debugger sends SIGCONT; in this case we need to retry the system
call).

Cancelling a graph also makes bt_graph_run() quit before the next
sink consuming iteration.

It is not allowed to run a canceled graph again. Once it is canceled,
it's dead for good.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>