ctf.fs source: recurse to find multiples CTF traces
One port per CTF data stream amongst all the traces that are found
from the user's starting path is created.

The traces are named so as to be able to replicate the original tree
by stripping their common prefix within slash boundaries.

For example: user passes `hello` as the starting path, the component
finds the following traces:

    hello/world/kilo
    hello/yes
    hello/ispep/zero/mini

These traces would be named:

    world/kilo
    yes
    ispep/zero/mini

Another example: user passes `hello` as the starting path, the
component finds the following trace:

    hello/world/kilo

In this case, the trace is named `kilo`.

I used realpath() to normalize the paths and make sure that there aren't
`.` or `..` in the trace's name. Opening a trace located directly in `/`
is not supported.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>