Remove bt_graph_add_component_as_sibling()
This is not used anywhere, and it's one more thing to test. It remains
in the archives in case we need it in the future. This was created in
the days when it was expected that components would duplicate themselves
to share resources instead of dynamically adding ports like today.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>